+++ ext/log-exporter/src/test/java/core/log/kafka/EventMessageHandlerTest.java
@@ -1,8 +1,11 @@
 package core.log.kafka;
 
+import core.framework.inject.Inject;
 import core.framework.kafka.Message;
 import core.framework.log.message.EventMessage;
 import core.framework.util.Files;
+import core.log.IntegrationTest;
+import core.log.domain.EventSchema;
 import core.log.service.ArchiveService;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -15,14 +18,17 @@ import java.util.List;
 /**
  * @author neo
  */
-class EventMessageHandlerTest {
+class EventMessageHandlerTest extends IntegrationTest {
     private EventMessageHandler handler;
+    @Inject
+    EventSchema schema;
 
     @BeforeEach
     void createEventMessageHandler() {
         handler = new EventMessageHandler();
         handler.archiveService = new ArchiveService();
         handler.archiveService.logDir = Files.tempDir();
+        handler.schema = schema;
     }
 
     @AfterEach
@@ -36,6 +42,10 @@ class EventMessageHandlerTest {
         message.date = Instant.parse("2022-11-07T00:00:00Z");
         message.id = "id";
         message.app = "app";
+        message.action = "action";
+        message.result = "OK";
+        message.elapsed = 1000L;
+        handler.handle(List.of(new Message<>("key", message)));
         handler.handle(List.of(new Message<>("key", message)));
     }
 }